feat: add DuckDB as optional storage backend (`memory-storage-duckdb`)
- Added `do-memory-storage-duckdb` crate implementing `StorageBackend` and `MonitoringStorageBackend`.
- Integrated DuckDB VSS extension for HNSW-based vector similarity search.
- Implemented columnar schema with support for episodic memory, patterns, agent metrics, relationships, and summaries.
- Added OLAP analytical queries for session windowing, temporal decay, and reward distribution.
- Added `run_analytics_query` for raw SQL analysis.
- **Enhanced Test Coverage**:
  - Expanded `integration_tests.rs` to cover all storage modules, trait methods, and analytical queries.
  - Reached >50% code coverage target for the new crate.
- **Hardened Implementation**:
  - Fixed multi-statement SQL execution.
  - Sanitized error messages to prevent data leaks.
  - Replaced silent fallbacks with proper error propagation.
- Verified code quality and addressed all PR feedback.
