+++ memory-mcp/src/bin/server_impl/oauth.rs
@@ -15,7 +15,7 @@ use do_memory_mcp::protocol::OAuthConfig;
 #[cfg(feature = "oauth")]
 use {
     super::types::AuthorizationResult,
-    jsonwebtoken::{DecodingKey, Validation, decode},
+    jsonwebtoken::{DecodingKey, Validation, dangerous::insecure_decode, decode},
     serde::{Deserialize, Serialize},
     tracing::{debug, warn},
 };
