+++ memory-storage-duckdb/src/storage/patterns.rs
@@ -53,7 +53,6 @@ impl DuckDbStorage {
         Ok(())
     }
 
-    #[allow(dead_code)]
     pub(crate) async fn load_patterns_internal(&self) -> Result<Vec<Pattern>> {
         let conn_arc = Arc::clone(&self.conn);
         let res = tokio::task::spawn_blocking(move || {
@@ -87,31 +86,30 @@ impl DuckDbStorage {
         .map_err(|e| Error::Storage(format!("Task join error: {e}")))??;
         Ok(res)
     }
+}
 
-    pub(crate) async fn get_pattern_internal(&self, id: uuid::Uuid) -> Result<Option<Pattern>> {
-        let conn_arc = Arc::clone(&self.conn);
-        let id_str = id.to_string();
-        let res = tokio::task::spawn_blocking(move || {
-            let conn = conn_arc.lock();
-            let mut stmt = conn
-                .prepare("SELECT CAST(pattern_data AS VARCHAR) FROM patterns WHERE pattern_id = ?")
-                .map_err(|e| Error::Storage(e.to_string()))?;
+pub(crate) async fn get_pattern_internal(&self, id: Uuid) -> Result<Option<Pattern>> {
+    let conn_arc = Arc::clone(&self.conn);
+    let res = tokio::task::spawn_blocking(move || {
+        let conn = conn_arc.lock();
+        let mut stmt = conn
+            .prepare("SELECT CAST(pattern_data AS VARCHAR) FROM patterns WHERE pattern_id = ?")
+            .map_err(|e| Error::Storage(e.to_string()))?;
 
-            let mut rows = stmt
-                .query(params![id_str])
-                .map_err(|e| Error::Storage(e.to_string()))?;
+        let mut rows = stmt
+            .query(params![id.to_string()])
+            .map_err(|e| Error::Storage(e.to_string()))?;
 
-            if let Some(row) = rows.next().map_err(|e| Error::Storage(e.to_string()))? {
-                let data_json: String = row.get(0).map_err(|e| Error::Storage(e.to_string()))?;
-                let pattern: Pattern = serde_json::from_str(&data_json)
-                    .map_err(|e| Error::Storage(format!("Failed to deserialize pattern: {e}")))?;
-                Ok::<Option<Pattern>, Error>(Some(pattern))
-            } else {
-                Ok::<Option<Pattern>, Error>(None)
-            }
-        })
-        .await
-        .map_err(|e| Error::Storage(format!("Task join error: {e}")))??;
-        Ok(res)
-    }
+        if let Some(row) = rows.next().map_err(|e| Error::Storage(e.to_string()))? {
+            let data_json: String = row.get(0).map_err(|e| Error::Storage(e.to_string()))?;
+            let pattern: Pattern = serde_json::from_str(&data_json)
+                .map_err(|e| Error::Storage(format!("Failed to deserialize pattern: {e}")))?;
+            Ok::<Option<Pattern>, Error>(Some(pattern))
+        } else {
+            Ok::<Option<Pattern>, Error>(None)
+        }
+    })
+    .await
+    .map_err(|e| Error::Storage(format!("Task join error: {e}")))??;
+    Ok(res)
 }
