+++ memory-storage-turso/src/storage/embeddings_internal.rs
@@ -227,6 +227,8 @@ impl TursoStorage {
 
     /// Delete an embedding
     pub(crate) async fn _delete_embedding_internal(&self, item_id: &str) -> Result<bool> {
+        let (conn, _conn_id) = self.get_connection_with_id().await?;
+
         #[cfg(feature = "turso_multi_dimension")]
         {
             return self.delete_embedding_dimension_aware(item_id).await;
@@ -234,7 +236,6 @@ impl TursoStorage {
 
         #[cfg(not(feature = "turso_multi_dimension"))]
         {
-            let (conn, _conn_id) = self.get_connection_with_id().await?;
             const SQL: &str = "DELETE FROM embeddings WHERE item_id = ?";
             let stmt = self
                 .prepared_cache
